#2: Use leader/owner entity for targeting purposes
This changes the Harmony postfix of `EAITarget.check` so that the entity used for targeting purposes is the entity's leader or owner, if any exists.

This applies to both the entity doing the targeting, and the entity being targeted.

Any other code changes are just refactoring to make the code cleaner and/or avoid unnecessary CPU cycles; they should not affect behavior.

Additionally, the "deployed" tag is added to the "HumanTags" in the config feature block. This will make deployed entities (junk sledge, junk turret, junk drone) use the same targeting check, so they no longer target NPCs hired by the player.